+++ arviz/stats/__init__.py
@@ -31,4 +31,5 @@ __all__ = [
     "wrap_xarray_ufunc",
     "reloo",
     "_calculate_ics",
+    "psens",
 ]