+++ arviz/stats/stats.py
@@ -30,6 +30,7 @@ from .density_utils import kde as _kde
 from .diagnostics import _mc_error, _multichain_statistics, ess
 from .stats_utils import ELPDData, _circular_standard_deviation, smooth_data
 from .stats_utils import get_log_likelihood as _get_log_likelihood
+from .stats_utils import get_log_prior as _get_log_prior
 from .stats_utils import logsumexp as _logsumexp
 from .stats_utils import make_ufunc as _make_ufunc
 from .stats_utils import stats_variance_2d as svar
@@ -51,6 +52,7 @@ __all__ = [
     "waic",
     "weight_predictions",
     "_calculate_ics",
+    "psens",
 ]
 
 
@@ -2113,3 +2115,219 @@ def weight_predictions(idatas, weights=None):
     )
 
     return weighted_samples
+
+
+def psens(
+    data,
+    *,
+    component="prior",
+    component_var_names=None,
+    component_coords=None,
+    var_names=None,
+    coords=None,
+    filter_vars=None,
+    delta=0.01,
+    dask_kwargs=None,
+):
+    """Compute power-scaling sensitivity diagnostic.
+
+    Power-scales the prior or likelihood and calculates how much the posterior is affected.
+
+    Parameters
+    ----------
+    data : obj
+        Any object that can be converted to an :class:`arviz.InferenceData` object.
+        Refer to documentation of :func:`arviz.convert_to_dataset` for details.
+        For ndarray: shape = (chain, draw).
+        For n-dimensional ndarray transform first to dataset with ``az.convert_to_dataset``.
+    component : {"prior", "likelihood"}, default "prior"
+        When `component` is "likelihood", the log likelihood values are retrieved
+        from the ``log_likelihood`` group as pointwise log likelihood and added
+        together. With "prior", the log prior values are retrieved from the
+        ``log_prior`` group.
+    component_var_names : str, optional
+        Name of the prior or log likelihood variables to use
+    component_coords : dict, optional
+        Coordinates defining a subset over the component element for which to
+        compute the prior sensitivity diagnostic.
+    var_names : list of str, optional
+        Names of posterior variables to include in the power scaling sensitivity diagnostic
+    coords : dict, optional
+        Coordinates defining a subset over the posterior. Only these variables will
+        be used when computing the prior sensitivity.
+    filter_vars: {None, "like", "regex"}, default None
+        If ``None`` (default), interpret var_names as the real variables names.
+        If "like", interpret var_names as substrings of the real variables names.
+        If "regex", interpret var_names as regular expressions on the real variables names.
+    delta : float
+        Value for finite difference derivative calculation.
+    dask_kwargs : dict, optional
+        Dask related kwargs passed to :func:`~arviz.wrap_xarray_ufunc`.
+
+    Returns
+    -------
+    xarray.Dataset
+        Returns dataset of power-scaling sensitivity diagnostic values.
+        Higher sensitivity values indicate greater sensitivity.
+        Prior sensitivity above 0.05 indicates informative prior.
+        Likelihood sensitivity below 0.05 indicates weak or nonin-formative likelihood.
+
+    Examples
+    --------
+    Compute the likelihood sensitivity for the non centered eight model:
+
+    .. ipython::
+
+        In [1]: import arviz as az
+           ...: data = az.load_arviz_data("non_centered_eight")
+           ...: az.psens(data, component="likelihood")
+
+    To compute the prior sensitivity, we need to first compute the log prior
+    at each posterior sample. In our case, we know mu has a normal prior :math:`N(0, 5)`,
+    tau is a half cauchy prior with scale/beta parameter 5,
+    and theta has a standard normal as prior.
+    We add this information to the ``log_prior`` group before computing powerscaling
+    check with ``psens``
+
+    .. ipython::
+
+        In [1]: from xarray_einstats.stats import XrContinuousRV
+           ...: from scipy.stats import norm, halfcauchy
+           ...: post = data.posterior
+           ...: log_prior = {
+           ...:     "mu": XrContinuousRV(norm, 0, 5).logpdf(post["mu"]),
+           ...:     "tau": XrContinuousRV(halfcauchy, scale=5).logpdf(post["tau"]),
+           ...:     "theta_t": XrContinuousRV(norm, 0, 1).logpdf(post["theta_t"]),
+           ...: }
+           ...: data.add_groups({"log_prior": log_prior})
+           ...: az.psens(data, component="prior")
+
+    Notes
+    -----
+    The diagnostic is computed by power-scaling the specified component (prior or likelihood)
+    and determining the degree to which the posterior changes as described in [1]_.
+    It uses Pareto-smoothed importance sampling to avoid refitting the model.
+
+    References
+    ----------
+    .. [1] Kallioinen et al, *Detecting and diagnosing prior and likelihood sensitivity with
+       power-scaling*, 2022, https://arxiv.org/abs/2107.14054
+
+    """
+    dataset = extract(data, var_names=var_names, filter_vars=filter_vars, group="posterior")
+    if coords is None:
+        dataset = dataset.sel(coords)
+
+    if component == "likelihood":
+        component_draws = _get_log_likelihood(data, var_name=component_var_names, single_var=False)
+    elif component == "prior":
+        component_draws = _get_log_prior(data, var_names=component_var_names)
+    else:
+        raise ValueError("Value for `component` argument not recognized")
+
+    component_draws = component_draws.stack(__sample__=("chain", "draw"))
+    if component_coords is None:
+        component_draws = component_draws.sel(component_coords)
+    if isinstance(component_draws, xr.DataArray):
+        component_draws = component_draws.to_dataset()
+    if len(component_draws.dims):
+        component_draws = component_draws.to_stacked_array(
+            "latent-obs_var", sample_dims=("__sample__",)
+        ).sum("latent-obs_var")
+    # from here component_draws is a 1d object with dimensions (sample,)
+
+    # calculate lower and upper alpha values
+    lower_alpha = 1 / (1 + delta)
+    upper_alpha = 1 + delta
+
+    # calculate importance sampling weights for lower and upper alpha power-scaling
+    lower_w = np.exp(_powerscale_lw(component_draws=component_draws, alpha=lower_alpha))
+    lower_w = lower_w / np.sum(lower_w)
+
+    upper_w = np.exp(_powerscale_lw(component_draws=component_draws, alpha=upper_alpha))
+    upper_w = upper_w / np.sum(upper_w)
+
+    ufunc_kwargs = {"n_dims": 1, "ravel": False}
+    func_kwargs = {"lower_weights": lower_w.values, "upper_weights": upper_w.values, "delta": delta}
+
+    # calculate the sensitivity diagnostic based on the importance weights and draws
+    return _wrap_xarray_ufunc(
+        _powerscale_sens,
+        dataset,
+        ufunc_kwargs=ufunc_kwargs,
+        func_kwargs=func_kwargs,
+        dask_kwargs=dask_kwargs,
+        input_core_dims=[["sample"]],
+    )
+
+
+def _powerscale_sens(draws, *, lower_weights=None, upper_weights=None, delta=0.01):
+    """
+    Calculate power-scaling sensitivity by finite difference
+    second derivative of CJS
+    """
+    lower_cjs = max(
+        _cjs_dist(draws=draws, weights=lower_weights),
+        _cjs_dist(draws=-1 * draws, weights=lower_weights),
+    )
+    upper_cjs = max(
+        _cjs_dist(draws=draws, weights=upper_weights),
+        _cjs_dist(draws=-1 * draws, weights=upper_weights),
+    )
+    logdiffsquare = 2 * np.log2(1 + delta)
+    grad = (lower_cjs + upper_cjs) / logdiffsquare
+
+    return grad
+
+
+def _powerscale_lw(alpha, component_draws):
+    """
+    Calculate log weights for power-scaling component by alpha.
+    """
+    log_weights = (alpha - 1) * component_draws
+    log_weights = psislw(log_weights)[0]
+
+    return log_weights
+
+
+def _cjs_dist(draws, weights):
+    """
+    Calculate the cumulative Jensen-Shannon distance between original draws and weighted draws.
+    """
+
+    # sort draws and weights
+    order = np.argsort(draws)
+    draws = draws[order]
+    weights = weights[order]
+
+    binwidth = np.diff(draws)
+
+    # ecdfs
+    cdf_p = np.linspace(1 / len(draws), 1 - 1 / len(draws), len(draws) - 1)
+    cdf_q = np.cumsum(weights / np.sum(weights))[:-1]
+
+    # integrals of ecdfs
+    cdf_p_int = np.dot(cdf_p, binwidth)
+    cdf_q_int = np.dot(cdf_q, binwidth)
+
+    # cjs calculation
+    pq_numer = np.log2(cdf_p, out=np.zeros_like(cdf_p), where=cdf_p != 0)
+    qp_numer = np.log2(cdf_q, out=np.zeros_like(cdf_q), where=cdf_q != 0)
+
+    denom = 0.5 * (cdf_p + cdf_q)
+    denom = np.log2(denom, out=np.zeros_like(denom), where=denom != 0)
+
+    cjs_pq = np.sum(binwidth * (cdf_p * (pq_numer - denom))) + 0.5 / np.log(2) * (
+        cdf_q_int - cdf_p_int
+    )
+
+    cjs_qp = np.sum(binwidth * (cdf_q * (qp_numer - denom))) + 0.5 / np.log(2) * (
+        cdf_p_int - cdf_q_int
+    )
+
+    cjs_pq = max(0, cjs_pq)
+    cjs_qp = max(0, cjs_qp)
+
+    bound = cdf_p_int + cdf_q_int
+
+    return np.sqrt((cjs_pq + cjs_qp) / bound)